fixup! refactor(ivy): move instructions
