+++ packages/core/src/render3/instructions/all.ts
@@ -28,7 +28,7 @@
 export * from './alloc_host_vars';
 export * from './change_detection';
 export * from './container';
-export * from './data';
+export * from './storage';
 export * from './di';
 export * from './element';
 export * from './element_container';
